better name for test data dir
